+++ Utils/Templates_writer/Network_template_writer.py
