+++ _03_Hardware-targets/GAP8/C_Parser.py
@@ -24,8 +24,8 @@ import os
 import numpy as np
 
 # DORY modules
-from _00_Parsers.Parser_HW_to_C import Parser_HW_to_C
-import _01_Utils.Templates_writer.Layer2D_template_writer as Layer2D_writer
+from Parsers.Parser_HW_to_C import Parser_HW_to_C
+import Utils.Templates_writer.Layer2D_template_writer as Layer2D_writer
 
 # Directory
 file_path = "/".join(os.path.realpath(__file__).split("/")[:-1])
